+++ packages/core/test/core.spec.ts
@@ -55,6 +55,17 @@ describe("Core functionality", () => {
     expect(button({ disabled: true })).toBe("bg-blue-100 opacity-50");
   });
 
+  it("Should not crash on empty variant", () => {
+    const button = classed("bg-blue-100", {
+      variants: {
+        test: {},
+      },
+    });
+
+    // @ts-expect-error test is not a valid variant
+    expect(button({ test: "lg" })).toBe("bg-blue-100");
+  });
+
   it("Should allow for composable boolean variants", () => {
     const BasicButton = classed("bg-blue-100", {
       variants: { loading: { true: "opacity-50" } },
@@ -264,5 +275,78 @@ describe("Composition", () => {
     });
 
     expect(button({ size: "sm" })).toBe("bg-blue-100 text-sm");
+  });
+});
+
+describe("With extra classes", () => {
+  it("Should add `class` to the list", () => {
+    const button = classed("bg-blue-100");
+
+    expect(button({ class: "text-white" })).toBe("bg-blue-100 text-white");
+  });
+
+  it("Should work with variants", () => {
+    const button = classed("bg-blue-100", {
+      variants: {
+        size: {
+          sm: "text-sm",
+          md: "text-md",
+        },
+      },
+    });
+
+    expect(button({ size: "sm", class: "text-white" })).toBe(
+      "bg-blue-100 text-sm text-white"
+    );
+  });
+
+  it("Should work with multiple classed functions", () => {
+    const button = classed("bg-blue-100", {
+      variants: {
+        size: {
+          sm: "text-sm",
+          md: "text-md",
+        },
+      },
+    });
+
+    const button2 = classed(button, "bg-red-100");
+
+    expect(
+      classIncludes(button2({ size: "sm", class: "text-white" }), [
+        "bg-blue-100",
+        "text-sm",
+        "bg-red-100",
+        "text-white",
+      ])
+    ).toBe(true);
+  });
+
+  it("Should work with className", () => {
+    const button = classed("bg-blue-100", {
+      variants: {
+        size: {
+          sm: "text-sm",
+          md: "text-md",
+        },
+      },
+    });
+
+    expect(button({ size: "sm", className: "text-white" })).toBe(
+      "bg-blue-100 text-sm text-white"
+    );
+  });
+
+  it("Should work with no variants added", () => {
+    const button = classed("bg-blue-100", {
+      variants: {
+        size: {
+          sm: "text-sm",
+          md: "text-md",
+        },
+      },
+    });
+
+    expect(button({ class: "text-white" })).toBe("bg-blue-100 text-white");
   });
 });